Disable scheduler PDB when scheduler is disabled
The helm chart doesn't check whether the scheduler itself is enabled
when deciding whether to configure the scheduler PDB.

This commit uses the value of Helm value `scheduler.enabled` to set the
default value of Helm value `scheduler.podDisruptionBudget.enabled`.

This should stop the spurious PodDisruptionBudgetAtLimit for the
scheduler PDB when the scheduler isn't deployed at all.